use unsigned big integer

+++ database/migrations/2016_10_22_105254_create_profiles_table.php
@@ -14,7 +14,7 @@ class CreateProfilesTable extends Migration
     public function up()
     {
         Schema::create('profiles', function (Blueprint $table) {
-            $table->unsignedInteger('profile_id')->unique();
+            $table->unsignedBigInteger('profile_id')->unique();
             $table->string('profile_gender');
             $table->string('profile_age_range');
             $table->string('profile_email')->unique();
