+++ database/migrations/2016_10_22_105321_create_works_table.php
@@ -14,7 +14,7 @@ class CreateWorksTable extends Migration
     public function up()
     {
         Schema::create('works', function (Blueprint $table) {
-            $table->unsignedInteger('work_id')->unique();
+            $table->unsignedBigInteger('work_id')->unique();
             $table->string('work_name');
             $table->timestamps();
         });
