+++ database/migrations/2016_10_22_105329_create_books_table.php
@@ -14,7 +14,7 @@ class CreateBooksTable extends Migration
     public function up()
     {
         Schema::create('books', function (Blueprint $table) {
-            $table->unsignedInteger('book_id')->unique();
+            $table->unsignedBigInteger('book_id')->unique();
             $table->string('book_name');
             $table->timestamps();
         });
